requirements.txt
streamlit
pytesseract
Pillow
pdf2image
pydantic
pydantic-settings
google-genai
ollama
